refactor: extract `isDatabaseConflictError` to plugin api
