+++ packages/backend-plugin-api/api-report.md
@@ -354,6 +354,9 @@ export interface IdentityService extends IdentityApi {}
 
 export { isChildPath };
 
+// @public
+export function isDatabaseConflictError(e: unknown): boolean;
+
 // @public (undocumented)
 export interface LifecycleService {
   addShutdownHook(
