+++ packages/backend-plugin-api/src/services/utilities/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Backstage Authors
+ * Copyright 2024 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +14,4 @@
  * limitations under the License.
  */
 
-export * from './definitions';
-export * from './system';
-export * from './utilities';
+export { isDatabaseConflictError } from './database';
